Update abacus_estimation.Rmd
Made a number of small changes throughout to formatting, notes, etc.. and to the display (e.g., to see all predictors and all sig. predictors). Also included some of the old growth curve stuff, but some of it broke (probably my fault...I replaced "female" with "age" to be consistent with our other models, and I probably did something dumb along the way...can fix later if we need this stuff)